+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/block/IColored.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.moonlight.api.block;
 
+import net.mehvahdjukaar.moonlight.api.set.BlocksColorAPI;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Item;
@@ -20,7 +21,10 @@ public interface IColored {
      * @return the associated item to this block or item with the given color. Null if it fails
      */
     @Nullable
-    Item changeItemColor(@Nullable DyeColor color);
+    @Deprecated(forRemoval = true)
+    default Item changeItemColor(@Nullable DyeColor color){
+        return BlocksColorAPI.changeColor(((ItemLike)this).asItem(), color);
+    };
 
     /**
      * If this kind of block can have a null color, similar to shulker boxes
@@ -29,12 +33,4 @@ public interface IColored {
         return false;
     }
 
-    /**
-     * casts the given object to this interface if it or its block are colored
-     */
-    static Optional<IColored> getOptional(ItemLike itemLike) {
-        if (itemLike instanceof IColored col) return Optional.of(col);
-        if (itemLike instanceof BlockItem bi && bi.getBlock() instanceof IColored col) return Optional.of(col);
-        return Optional.empty();
-    }
 }