+++ forge/src/main/java/net/mehvahdjukaar/moonlight/core/misc/forge/ModLootConditions.java
@@ -15,9 +15,12 @@ import net.minecraftforge.common.crafting.CraftingHelper;
 import net.minecraftforge.common.crafting.conditions.ICondition;
 
 import org.jetbrains.annotations.NotNull;
+
+import javax.annotation.Nonnull;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
+import java.util.regex.Pattern;
 
 public class ModLootConditions {
 
@@ -26,6 +29,8 @@ public class ModLootConditions {
 
     public static final Supplier<LootItemConditionType> ICONDITION_LOOT_CONDITION = RegHelper.register(Moonlight.res("iconditions"),
             () -> new LootItemConditionType(new IConditionLootCondition.ConditionSerializer()), Registries.LOOT_CONDITION_TYPE);
+    public static final Supplier<LootItemConditionType> PATTERN_MATCH_CONDITION = RegHelper.register(Moonlight.res("loot_table_id_patter"),
+            () -> new LootItemConditionType(new PatternMatchCondition.ConditionSerializer()), Registries.LOOT_CONDITION_TYPE);
 
     public record IConditionLootCondition(List<ICondition> conditions) implements LootItemCondition {
         @Override
@@ -63,6 +68,48 @@ public class ModLootConditions {
                     l.add(CraftingHelper.getCondition(c.getAsJsonObject()));
                 }
                 return new IConditionLootCondition(l);
+            }
+        }
+    }
+
+
+    public record PatternMatchCondition(List<Pattern> patterns) implements LootItemCondition {
+
+        @Override
+        public boolean test(LootContext lootContext) {
+            String id = lootContext.getQueriedLootTableId().toString();
+            for (var p : patterns) {
+                if (id.equals(p.pattern())) return true;
+                if (p.matcher(id).find()) return true;
+            }
+            return false;
+        }
+
+        @Nonnull
+        @Override
+        public LootItemConditionType getType() {
+            return PATTERN_MATCH_CONDITION.get();
+        }
+
+        public record ConditionSerializer() implements Serializer<PatternMatchCondition> {
+            @Override
+            public void serialize(@Nonnull JsonObject json, @Nonnull PatternMatchCondition value, @Nonnull JsonSerializationContext context) {
+                JsonArray ja = new JsonArray();
+                for (var c : value.patterns) {
+                    ja.add(c.pattern());
+                }
+                json.add("matches", ja);
+            }
+
+            @Nonnull
+            @Override
+            public PatternMatchCondition deserialize(@Nonnull JsonObject json, @Nonnull JsonDeserializationContext context) {
+                var ja = GsonHelper.getAsJsonArray(json, "matches");
+                List<Pattern> l = new ArrayList<>();
+                for (var c : ja) {
+                    l.add(Pattern.compile(c.getAsString()));
+                }
+                return new PatternMatchCondition(l);
             }
         }
     }