+++ forge/src/main/java/net/mehvahdjukaar/moonlight/core/misc/forge/ModLootModifiers.java
@@ -5,8 +5,10 @@ import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.mehvahdjukaar.moonlight.core.Moonlight;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.LootContext;
+import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
 import net.minecraftforge.common.loot.IGlobalLootModifier;
 import net.minecraftforge.common.loot.LootModifier;
@@ -32,6 +34,9 @@ public class ModLootModifiers {
 
     public static final RegistryObject<Codec<? extends IGlobalLootModifier>> REPLACE_ITEM_GLM =
             LOOT_MODIFIERS.register("replace_item", ReplaceItemModifier.CODEC);
+
+    public static final RegistryObject<Codec<? extends IGlobalLootModifier>> ADD_TABLE =
+            LOOT_MODIFIERS.register("add_loot_table", AddTableModifier.CODEC);
 
     public static class AddItemModifier extends LootModifier {
 
@@ -95,6 +100,35 @@ public class ModLootModifiers {
             if (!generatedLoot.isEmpty()) {
                 generatedLoot.set(0, itemStack.copy());
             }
+            return generatedLoot;
+        }
+
+        @Override
+        public Codec<? extends IGlobalLootModifier> codec() {
+            return CODEC.get();
+        }
+    }
+
+
+    public static class AddTableModifier extends LootModifier {
+
+        public static final Supplier<Codec<AddTableModifier>> CODEC = Suppliers.memoize(() -> RecordCodecBuilder.create(inst -> codecStart(inst).and(
+                        ResourceLocation.CODEC.fieldOf("loot_table").forGetter(m -> m.injectTableId)
+                ).apply(inst, AddTableModifier::new)
+        ));
+
+        private final ResourceLocation injectTableId;
+
+        protected AddTableModifier(LootItemCondition[] conditionsIn, ResourceLocation addedItemStack) {
+            super(conditionsIn);
+            this.injectTableId = addedItemStack;
+        }
+
+        @NotNull
+        @Override
+        protected ObjectArrayList<ItemStack> doApply(ObjectArrayList<ItemStack> generatedLoot, LootContext context) {
+            LootTable lootTable = context.getLevel().getServer().getLootData().getLootTable(injectTableId);
+            lootTable.getRandomItems(context, generatedLoot::add);
             return generatedLoot;
         }